refactor(sessions): keep lifecycle routing within baselines
